requirements.txt
fastapi==0.109.2
uvicorn==0.27.1 
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
supabase>=2.3.0
chunkr-ai>=0.0.50 